requirements.txt
Flask==3.0.2
pymongo==4.6.2
python-dotenv==1.0.1
mysql-connector-python==8.3.0
openai==1.12.0